Resolve clippy warnings
