+++ src/model/event.rs
@@ -1768,7 +1768,7 @@ impl<'de> Deserialize<'de> for GatewayEvent {
                     .map_err(DeError::custom)?;
                 let payload = map
                     .remove("d")
-                    .ok_or(Error::Decode("expected gateway event d", Value::from(map)))
+                    .ok_or_else(|| Error::Decode("expected gateway event d", Value::from(map)))
                     .map_err(DeError::custom)?;
 
                 let x = match deserialize_event_with_type(kind.clone(), payload) {
